+++ resources/views/tailwind/components/table/table.blade.php
@@ -1,3 +1,5 @@
+@props(['useHeaderAsFooter' => false, 'customFooter' => false])
+
 <div class="align-middle min-w-full overflow-x-auto shadow overflow-hidden rounded-none md:rounded-lg">
     <table {{ $attributes->except('wire:sortable') }} class="min-w-full divide-y divide-gray-200 dark:divide-none">
         <thead>
@@ -9,5 +11,17 @@
         <tbody {{ $attributes->only('wire:sortable') }} class="bg-white divide-y divide-gray-200 dark:divide-none">
             {{ $body }}
         </tbody>
+
+        @if ($useHeaderAsFooter || $customFooter)
+            <tfoot>
+                @if ($useHeaderAsFooter)
+                    <tr>
+                        {{ $head }}
+                    </tr>
+                @elseif($customFooter)
+                    {{ $foot }}
+                @endif
+            </tfoot>
+        @endif
     </table>
 </div>